Update Vagrantfile to ubuntu/xenial64.
The version of Redis that ships with Ubuntu Trusty is too old to support
the HyperLogLog commands.  I guess we could add a custom apt repo for
this, but it seems a lot easier to just use the latest Ubuntu LTS
release (i.e. xenial), which has a recent-enough Redis in its archives.