German translation issue on landingpage CTA (#8865)
There is a German translation issue on the landingpage's CTA showing number of users and status posted on the instance.

But as it is right now a verb is missing for the number of status posted. But adding there the right verb like "Zuhause für X Benutzer die Y Status veröffentlichten" the text at the bottom of the numbers is getting a bit too big.

That's the reason I changed wording to "Zuhause für X Benutzer mit Y Statusmeldungen" which seems shorter and more in balance regarding top and bottom text on the numbers.